refactor(wallet): move routing into its own module
