+++ wallet/src/actors/controller.rs
@@ -6,10 +6,8 @@ use std::net;
 use std::path::PathBuf;
 
 use actix::prelude::*;
-use futures::future;
 use jsonrpc_core as rpc;
 use jsonrpc_pubsub as pubsub;
-use serde_json::{self as json, json};
 
 use super::App;
 use crate::api;
@@ -26,59 +24,6 @@ impl Controller {
     pub fn build() -> ControllerBuilder {
         ControllerBuilder::new()
     }
-}
-
-/// Helper macro to add multiple JSON-RPC methods at once
-macro_rules! routes {
-    ($io:expr, $app:expr $(,)?) => {};
-    ($io:expr, $app:expr, ($method_jsonrpc:expr, $actor_msg:ty $(,)?), $($args:tt)*) => {
-        {
-            let app_addr = $app.clone();
-            $io.add_method($method_jsonrpc, move |params: rpc::Params| {
-                log::debug!("Handling request for method: {}", $method_jsonrpc);
-                let addr = app_addr.clone();
-                // Try to parse the request params into the actor message
-                future::result(params.parse::<$actor_msg>())
-                    .and_then(move |msg| {
-                        // Then send the parsed message to the actor
-                        addr.send(msg)
-                            .map_err(error::Error::Mailbox)
-                            .flatten()
-                            .and_then(
-                                |x|
-                                future::result(json::to_value(x)).map_err(error::Error::Serialization)
-                            )
-                            .map_err(|err| error::ApiError::Execution(err).into())
-                    })
-            });
-        }
-        routes!($io, $app, $($args)*);
-    };
-}
-
-/// Macro to add multiple JSON-RPC methods that forward the request to the Node at once
-macro_rules! forwarded_routes {
-    ($io:expr, $app:expr $(,)?) => {};
-    ($io:expr, $app:expr, $method_jsonrpc:expr, $($args:tt)*) => {
-        {
-            let app_addr = $app.clone();
-            $io.add_method($method_jsonrpc, move |params: rpc::Params| {
-                log::debug!("Forwarding request for method: {}", $method_jsonrpc);
-                let msg = api::ForwardRequest {
-                    method: $method_jsonrpc.to_string(),
-                    params
-                };
-                app_addr.send(msg)
-                    .map_err(error::Error::Mailbox)
-                    .flatten()
-                    .and_then(|x| {
-                        future::result(json::to_value(x)).map_err(error::Error::Serialization)
-                    })
-                    .map_err(|err| error::ApiError::Execution(err).into())
-            });
-        }
-        forwarded_routes!($io, $app, $($args)*);
-    };
 }
 
 /// Controller builder used to set optional parameters using the builder-pattern.
@@ -133,50 +78,10 @@ impl ControllerBuilder {
             .node_url(self.node_url)
             .db_path(self.db_path)
             .start()?;
+
         let mut handler = pubsub::PubSubHandler::new(rpc::MetaIoHandler::default());
 
-        handler.add_subscription(
-            "notifications",
-            ("subscribeNotifications", {
-                let addr = app.clone();
-                move |_, _, subscriber| addr.do_send(api::SubscribeRequest(subscriber))
-            }),
-            ("unsubscribeNotifications", {
-                let addr = app.clone();
-                move |id, _| {
-                    addr.send(api::UnsubscribeRequest(id))
-                        .map_err(error::Error::Mailbox)
-                        .and_then(|_| future::ok(json!({"status": "ok"})))
-                        .map_err(|err| error::ApiError::Execution(err).into())
-                }
-            }),
-        );
-
-        forwarded_routes!(
-            handler,
-            app,
-            "getBlock",
-            "getBlockChain",
-            "getOutput",
-            "inventory",
-        );
-
-        routes!(
-            handler,
-            app,
-            ("getWalletInfos", api::WalletInfosRequest),
-            ("createMnemonics", api::CreateMnemonicsRequest),
-            ("importSeed", api::ImportSeedRequest),
-            ("createWallet", api::CreateWalletRequest),
-            ("lockWallet", api::LockWalletRequest),
-            ("unlockWallet", api::UnlockWalletRequest),
-            ("getTransactions", api::GetTransactionsRequest),
-            ("sendVTT", api::SendVttRequest),
-            ("generateAddress", api::GenerateAddressRequest),
-            ("createDataRequest", api::CreateDataReqRequest),
-            ("runRadRequest", api::RunRadReqRequest),
-            ("sendDataRequest", api::SendDataReqRequest),
-        );
+        api::connect_routes(&mut handler, app.clone());
 
         let server = Server::build()
             .handler(handler)
