+++ wallet/src/api/endpoints/create_mnemonics.rs
