+++ wallet/src/api/endpoints/create_wallet.rs
