+++ wallet/src/api/endpoints/generate_address.rs
