+++ wallet/src/api/endpoints/get_wallet_infos.rs
