+++ wallet/src/api/endpoints/lock_wallet.rs
