+++ wallet/src/api/endpoints/responses.rs
