+++ wallet/src/api/mod.rs
@@ -1,32 +1,5 @@
-pub mod create_data_req;
-pub mod create_mnemonics;
-pub mod create_wallet;
-pub mod forward;
-pub mod generate_address;
-pub mod get_transactions;
-pub mod get_wallet_infos;
-pub mod import_seed;
-pub mod lock_wallet;
-pub mod run_rad_req;
-pub mod send_data_req;
-pub mod send_vtt;
-pub mod subscribe;
-pub mod unlock_wallet;
-pub mod unsubscribe;
+pub mod endpoints;
+pub mod routes;
 
-pub use create_mnemonics::*;
-pub use create_wallet::*;
-pub use forward::*;
-pub use get_transactions::*;
-pub use get_wallet_infos::*;
-pub use import_seed::*;
-pub use lock_wallet::*;
-pub use subscribe::*;
-pub use unlock_wallet::*;
-pub use unsubscribe::*;
-
-pub use create_data_req::*;
-pub use generate_address::*;
-pub use run_rad_req::*;
-pub use send_data_req::*;
-pub use send_vtt::*;
+pub use endpoints::*;
+pub use routes::*;
